+++ app/models/user.ts
@@ -10,7 +10,7 @@ import OsfModel from './osf-model';
 import Region from './region';
 import Registration from './registration';
 import UserEmail from './user-email';
-import UserSettings from './user-settings';
+import UserSetting from './user-setting';
 
 /**
  * @module ember-osf-web
@@ -77,7 +77,7 @@ export default class User extends OsfModel.extend(Validations) {
     @attr('object') social!: {};
 
     @belongsTo('region') defaultRegion!: DS.PromiseObject<Region> & Region;
-    @belongsTo('user-settings') settings!: DS.PromiseObject<UserSettings> & UserSettings;
+    @belongsTo('user-setting', { inverse: 'user' }) settings!: DS.PromiseObject<UserSetting> & UserSetting;
 
     @hasMany('node') nodes!: DS.PromiseManyArray<Node>;
     @hasMany('contributor') contributors!: DS.PromiseManyArray<Contributor>;