+++ types/ember-cli-mirage/index.d.ts
@@ -96,7 +96,7 @@ export interface Request {
 export interface HandlerContext {
     request: Request;
     serialize(modelOrCollection: ModelInstance | ModelInstance[] | ModelClass, serializerName?: string): any;
-    normalizedRequestAttrs(): any;
+    normalizedRequestAttrs(model?: string): any;
 }
 interface HandlerObject {
     [k: string]: any;