+++ CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
     - `new-project-navigation-modal` component
 - Models:
     - `user-email`
-    - `user-settings`
+    - `user-setting`
 - Routes:
     - `settings.account`
     - `settings.profile` - redirects to `settings.profile.name`
@@ -43,6 +43,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 - Addons:
+    - `ember-cli-mirage@0.4.10`
     - `ember-code-snippet@2.3.1`
 - Components:
     - `node-navbar` - Choose links to display with the same logic as legacy